+++ src/background/seelen_weg/handler.rs
@@ -1,7 +1,7 @@
 use std::{ffi::OsStr, path::PathBuf};
 
 use image::ImageFormat;
-use seelen_core::state::{PinnedWegItemData, WegItem, WegItemSubtype, WegItems};
+use seelen_core::state::{PinnedWegItemData, RelaunchArguments, WegItem, WegItemSubtype, WegItems};
 use tauri::Emitter;
 use tauri_plugin_shell::ShellExt;
 
@@ -109,6 +109,7 @@ pub fn weg_toggle_window_state(hwnd: isize, was_focused: bool) -> Result<()> {
     Ok(())
 }
 
+#[allow(deprecated)]
 #[tauri::command(async)]
 pub fn weg_pin_item(path: PathBuf) -> Result<()> {
     let display_name = if let Some(name) = path.file_name() {
@@ -133,8 +134,10 @@ pub fn weg_pin_item(path: PathBuf) -> Result<()> {
         display_name,
         path: path.clone(),
         is_dir: false,
+        relaunch_command: None,
+        relaunch_program: path.to_string_lossy().to_string(),
+        relaunch_args: None,
         relaunch_in: None,
-        relaunch_command: path.to_string_lossy().to_string(),
         windows: vec![],
         pin_disabled: false,
     };
@@ -143,11 +146,11 @@ pub fn weg_pin_item(path: PathBuf) -> Result<()> {
         data.umid = WindowsApi::get_file_umid(&path).ok();
         let (program, arguments) = WindowsApi::resolve_lnk_target(&path)?;
         data.is_dir = program.is_dir();
-        data.relaunch_command = format!(
-            "\"{}\" {}",
-            program.to_string_lossy(),
-            arguments.to_string_lossy()
-        );
+        data.relaunch_program = program.to_string_lossy().to_string(); //
+        data.relaunch_args = Some(RelaunchArguments::String(
+            arguments.to_string_lossy().to_string(),
+        ));
+
         if program.extension() == Some(OsStr::new("exe")) {
             data.subtype = WegItemSubtype::App;
         }